Fix phys_page macro
In RISC-V, PTEs have 10 bits of info on the right hand side, so to
convert a PTE to a physical address you shift right by 10 bits and then
left by 12.

This commit updates the phys_page macro to work for RISC-V and replaces
several instances where we manually performed the shift.