Revert changes in BootAux and use 2MB huge pages for guest kernel
Reverts changes from the previous commit affecting BootAux. We now
use 2MB huge pages again to map the kernel image in guest physical
mem.

This commit also adds a safeguard in vm.c to ensure that the physical
pages backing the kernel image are suitable for mapping with 2MB huge
pages. This means the 4k pages comprising the image are either 2MB
aligned or contiguous with a 2MB aligned region in the kernel image.

This patch accompanies a patch to kvmtool that causes it to map the
kernel image using hugepages. Note that the kernel boot string passed
to qemu must pre-allocate hugepages for kvmtool to use.